+++ crates/re_viewer/src/ui/selection_panel.rs
@@ -1,9 +1,11 @@
 use re_data_store::{log_db::LogDb, ObjectProps};
 use re_log_types::LogMsg;
 
-use crate::{data_ui::*, ui::Blueprint, Preview, Selection, ViewerContext};
-
-use super::SpaceView;
+use crate::{
+    data_ui::*,
+    ui::{view_3d, view_text, Blueprint, SpaceView},
+    Preview, Selection, ViewerContext,
+};
 
 // ---
 
@@ -240,16 +242,21 @@ fn ui_space_view(ctx: &mut ViewerContext<'_>, ui: &mut egui::Ui, space_view: &mu
     use super::space_view::ViewCategory;
     match space_view.category {
         ViewCategory::ThreeD => {
-            ui.label("3D view.");
-            super::view_3d::show_settings_ui(ctx, ui, &mut space_view.view_state.state_3d);
+            ui.strong("3D view");
+            view_3d::show_settings_ui(ctx, ui, &mut space_view.view_state.state_3d);
         }
         ViewCategory::Tensor => {
             if let Some(state_tensor) = &mut space_view.view_state.state_tensor {
-                ui.label("Tensor view.");
+                ui.strong("Tensor view");
                 state_tensor.ui(ui);
             }
         }
-        ViewCategory::TwoD | ViewCategory::Text | ViewCategory::Plot => {}
+        ViewCategory::Text => {
+            ui.strong("Text view");
+            ui.add_space(4.0);
+            view_text::text_filters_ui(ui, &mut space_view.view_state.state_text);
+        }
+        ViewCategory::TwoD | ViewCategory::Plot => {}
     }
 }