+++ crates/re_viewer/src/ui/view_text/scene.rs
@@ -3,6 +3,8 @@ use re_log_types::{IndexHash, MsgId, ObjectType};
 
 use crate::{ui::SceneQuery, ViewerContext};
 
+use super::ui::ViewTextFilters;
+
 // ---
 
 /// A single text entry.
@@ -26,19 +28,34 @@ pub struct SceneText {
 
 impl SceneText {
     /// Loads all text objects into the scene according to the given query.
-    pub(crate) fn load_objects(&mut self, ctx: &ViewerContext<'_>, query: &SceneQuery<'_>) {
+    pub(crate) fn load_objects(
+        &mut self,
+        ctx: &ViewerContext<'_>,
+        query: &SceneQuery<'_>,
+        filters: &ViewTextFilters,
+    ) {
         crate::profile_function!();
 
-        self.load_text_entries(ctx, query);
+        self.load_text_entries(ctx, query, filters);
     }
 
-    fn load_text_entries(&mut self, ctx: &ViewerContext<'_>, query: &SceneQuery<'_>) {
+    fn load_text_entries(
+        &mut self,
+        ctx: &ViewerContext<'_>,
+        query: &SceneQuery<'_>,
+        filters: &ViewTextFilters,
+    ) {
         crate::profile_function!();
 
-        let text_entries = query
+        query
             .iter_object_stores(ctx.log_db, &[ObjectType::TextEntry])
-            .flat_map(|(_obj_type, obj_path, _time_query, obj_store)| {
-                let mut batch = Vec::new();
+            .for_each(|(_obj_type, obj_path, _time_query, obj_store)| {
+                // Early filtering: if we're not showing it the view, there isn't much point
+                // in querying it to begin with... at least for now.
+                if !filters.is_obj_path_visible(obj_path) {
+                    return;
+                }
+
                 // TODO(cmc): We're cloning full strings here, which is very much a bad idea.
                 // We need to change the internal storage so that we store ref-counted strings
                 // rather than plain strings.
@@ -57,26 +74,28 @@ impl SceneText {
                      body: &String,
                      level: Option<&String>,
                      color: Option<&[u8; 4]>| {
-                        batch.push(TextEntry {
-                            msg_id: *msg_id,
-                            obj_path: obj_path.clone(),
-                            time,
-                            color: color.copied(),
-                            level: level.map(ToOwned::to_owned),
-                            body: body.clone(),
-                        });
+                        // Early filtering once more, see above.
+                        let is_visible = level
+                            .as_ref()
+                            .map_or(true, |lvl| filters.is_log_level_visible(lvl));
+
+                        if is_visible {
+                            self.text_entries.push(TextEntry {
+                                msg_id: *msg_id,
+                                obj_path: obj_path.clone(),
+                                time,
+                                color: color.copied(),
+                                level: level.map(ToOwned::to_owned),
+                                body: body.clone(),
+                            });
+                        }
                     },
                 );
-                batch
             });
 
-        self.text_entries.extend(text_entries);
-
         // We want to show the log messages in order.
-        // The most important order is the the `time` for whatever
-        // timeline we are on.
-        // For a tie-breaker, we use MsgId as that is
-        // ordered by a high-resolution wall-time.
+        // The most important order is the the `time` for whatever timeline we are on.
+        // For a tie-breaker, we use MsgId as that is ordered by a high-resolution wall-time.
         crate::profile_scope!("sort");
         self.text_entries
             .sort_by_key(|entry| (entry.time, entry.msg_id));